line chart: trasnfer data with 64 bit precision
We premature optimized thinking that we should allocate less memory
using Float32Array where possible instead of Float64Array. The thing
is, with 32 bit precision, it is not possible to represent the date
serialized to integer.

This changes the data transfer layer only. We will diagnose and change
other places where we use Float32Array as needed (for instance, WebGL
uses 32 bit precision and threejs internally uses Float32Array a lot.
Having more precision does not help in all cases).

+++ tensorboard/webapp/widgets/line_chart_v2/lib/worker/compact_data_series.ts
@@ -34,7 +34,7 @@ export function compactDataSeries(
     return len + data.points.length;
   }, 0);
   let seriesIndex = 0;
-  const flattenedSeries = new Float32Array(totalLength * 2);
+  const flattenedSeries = new Float64Array(totalLength * 2);
   const idsAndLengths: Array<{id: string; length: number}> = [];
 
   for (const series of dataSeriesArr) {
@@ -57,7 +57,7 @@ export function decompactDataSeries(
   compactDataSeries: CompactDataSeries
 ): DataSeries[] {
   const {flattenedSeries, idsAndLengths} = compactDataSeries;
-  const rawData = new Float32Array(flattenedSeries);
+  const rawData = new Float64Array(flattenedSeries);
   const data: DataSeries[] = [];
 
   if (rawData.length % 2 !== 0) {
